improv: refactor structure to fit with utilities
