+++ aws_lambda_powertools/utilities/validation/envelopes/__init__.py
@@ -1,0 +1,9 @@
+from .base import UserEnvelope
+from .dynamodb import DynamoDBEnvelope
+from .event_bridge import EventBridgeEnvelope
+
+__all__ = [
+    "UserEnvelope",
+    "DynamoDBEnvelope",
+    "EventBridgeEnvelope",
+]
