+++ aws_lambda_powertools/utilities/validation/schemas/sns.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class SnsSchema(BaseModel):
+    todo: str
