+++ aws_lambda_powertools/utilities/validation/schemas/sqs.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class SqsSchema(BaseModel):
+    todo: str
