--- aws_lambda_powertools/validation/__init__.py
@@ -1,2 +1,0 @@
-"""Validation utility
-"""
